Fixed issues suggested by bjedwards
